+++ lib/assets/javascripts/cartodb3/data/column-model.js
