+++ lib/assets/javascripts/cartodb3/data/columns-collection.js
