+++ lib/assets/javascripts/cartodb3/data/layer-definition-model.js
