+++ lib/assets/javascripts/cartodb3/data/layer-definitions-collection.js
