+++ lib/assets/javascripts/cartodb3/data/table-model.js
