+++ lib/assets/javascripts/cartodb3/data/tables-collection.js
