+++ lib/assets/javascripts/cartodb3/data/user-model.js
