+++ lib/assets/javascripts/cartodb3/data/visualization-model.js
