+++ lib/assets/javascripts/cartodb3/data/widget-definition-model.js
