+++ lib/assets/javascripts/cartodb3/editor.js
@@ -1,11 +1,11 @@
 var _ = require('underscore');
 var cdb = require('cartodb-deep-insights.js');
-var WidgetDefinitionModel = require('./data-models/widget-definition-model');
-var LayerDefinitionModel = require('./data-models/layer-definition-model');
-var VisualizationModel = require('./data-models/visualization-model');
+var WidgetDefinitionModel = require('./data/widget-definition-model');
+var LayerDefinitionModel = require('./data/layer-definition-model');
+var VisualizationModel = require('./data/visualization-model');
 var EditorWidgetsView = require('./editor/editor-widgets-view');
 var Polyglot = require('node-polyglot');
-var User = require('./data-models/user-model');
+var User = require('./data/user-model');
 var Config = require('./top-level-apis/config');
 
 // JSON data passed from entry point (editor/visualizations/show.html):
