+++ lib/assets/test/spec/cartodb3/data/table-model.spec.js
@@ -1,8 +1,8 @@
 var cdb = require('cartodb-deep-insights.js');
 var Config = require('../../../../javascripts/cartodb3/top-level-apis/config');
-var TableModel = require('../../../../javascripts/cartodb3/data-models/table-model');
+var TableModel = require('../../../../javascripts/cartodb3/data/table-model');
 
-describe('data-models/table-model', function () {
+describe('data/table-model', function () {
   beforeEach(function () {
     cdb.config = new Config();
     this.model = new TableModel({
