+++ lib/assets/test/spec/cartodb3/data/tables-collection.spec.js
@@ -1,7 +1,7 @@
 var Backbone = require('backbone');
-var TablesCollection = require('../../../../javascripts/cartodb3/data-models/tables-collection');
+var TablesCollection = require('../../../../javascripts/cartodb3/data/tables-collection');
 
-describe('data-models/tables-collection', function () {
+describe('data/tables-collection', function () {
   beforeEach(function () {
     this.collection = new TablesCollection([], {
       baseUrl: '/user/pepe'
