+++ lib/assets/test/spec/cartodb3/data/user-model.spec.js
@@ -1,8 +1,8 @@
 var cdb = require('cartodb-deep-insights.js');
 var Config = require('../../../../javascripts/cartodb3/top-level-apis/config');
-var UserModel = require('../../../../javascripts/cartodb3/data-models/user-model');
+var UserModel = require('../../../../javascripts/cartodb3/data/user-model');
 
-describe('data-models/user-model', function () {
+describe('data/user-model', function () {
   beforeEach(function () {
     cdb.config = new Config();
     this.model = new UserModel({
