+++ lib/assets/test/spec/cartodb3/data/visualization-model.spec.js
@@ -1,4 +1,4 @@
-var VisualizationModel = require('../../../../javascripts/cartodb3/data-models/visualization-model');
+var VisualizationModel = require('../../../../javascripts/cartodb3/data/visualization-model');
 
 describe('visualization-model', function () {
   beforeEach(function () {
