+++ lib/assets/test/spec/cartodb3/data/widget-definition-model.spec.js
@@ -1,6 +1,6 @@
 var cdb = require('cartodb-deep-insights.js');
-var WidgetDefinitionModel = require('../../../../javascripts/cartodb3/data-models/widget-definition-model');
-var LayerDefinitionModel = require('../../../../javascripts/cartodb3/data-models/layer-definition-model');
+var WidgetDefinitionModel = require('../../../../javascripts/cartodb3/data/widget-definition-model');
+var LayerDefinitionModel = require('../../../../javascripts/cartodb3/data/layer-definition-model');
 
 describe('widget-definition-model', function () {
   beforeEach(function () {
