+++ lib/assets/test/spec/cartodb3/editor/editor-widget-view.spec.js
@@ -1,5 +1,5 @@
 var EditorWidgetView = require('../../../../javascripts/cartodb3/editor/editor-widget-view');
-var WidgetDefinitionModel = require('../../../../javascripts/cartodb3/data-models/widget-definition-model');
+var WidgetDefinitionModel = require('../../../../javascripts/cartodb3/data/widget-definition-model');
 var cdb = require('cartodb-deep-insights.js');
 
 describe('editor/editor-widget-view', function () {
