+++ lib/assets/test/spec/cartodb3/editor/editor-widgets-view.spec.js
@@ -1,6 +1,6 @@
 var EditorWidgetsView = require('../../../../javascripts/cartodb3/editor/editor-widgets-view');
-var VisualizationModel = require('../../../../javascripts/cartodb3/data-models/visualization-model');
-var WidgetDefinitionModel = require('../../../../javascripts/cartodb3/data-models/widget-definition-model');
+var VisualizationModel = require('../../../../javascripts/cartodb3/data/visualization-model');
+var WidgetDefinitionModel = require('../../../../javascripts/cartodb3/data/widget-definition-model');
 var cdb = require('cartodb-deep-insights.js');
 
 describe('editor/editor-widgets-view', function () {
